kernel: save ip before store memory access
For load memory access, instrument trace function before the
memory access as we don't need to wait for the memory access
completion to get value. Also get ip in trace function.

For store memory access, instrument a `saveip` function to get
the memory access instruction access through macro `_RET_IP_`.
Intrument the trace function after the memory access. Notice that
there could be interrupt in the middle of the two functions, so use
an array (or stack) to keep instruction address (similar as saving
context before the interrupt).

Now for scalar memory access we can always get the correct ip and
value. Sometimes the ip is not accurate: e.g. consider a `jmp`
instruction that jumps to a memory store access. The compiler will
instrument the `saveip` before this `jmp` instruction. This will result
in saving the ip address of `jmp` instruction.

For memory region access, the instruction address is not accurate.