gcc: instrument before & after call instruction;
Previously we instrument before call and ret instruction. This lead to inconsistent call stack since we selectively instrument modules. Now we only instrument the call instruction.

For each call instruction, generate a unique ID for debugging usage. The ID will be saved by instrumentation function so we can check if the call and return traces match.